[CSLA-62] Use YAML config file instead of optparse-applicative
The long command line arguments cause problems in Windows.
