+++ cardano-sl.cabal
@@ -1077,13 +1077,16 @@ executable cardano-keygen
 executable cardano-launcher
   hs-source-dirs:      src/launcher
   main-is:             Main.hs
+  other-modules:       Launcher
   build-depends:       async
                      , base
                      , optparse-applicative
                      , cardano-report-server >= 0.1.0
                      , log-warper >= 0.4.3
                      , cardano-sl
+                     , bytestring
                      , directory
+                     , filepath
                      , process
                      , system-filepath
                      , text
@@ -1091,6 +1094,7 @@ executable cardano-launcher
                      , universum
                      , wreq
                      , lens
+                     , yaml
   default-language:    Haskell2010
   ghc-options:         -threaded
                        -Wall
@@ -1098,8 +1102,6 @@ executable cardano-launcher
   default-extensions:  OverloadedStrings
                        RecordWildCards
                        TupleSections
-  if os(windows)
-    LD-Options:     -mwindows
 
 test-suite cardano-test
   main-is:             Test.hs
